integrating add edit activites into one
