+++ MoodSpace/app/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".EditMood"></activity>
         <activity android:name=".ListActivity" />
         <activity android:name=".LoginActivity">
             <intent-filter>
@@ -20,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".AddMood"
+        <activity android:name=".AddEditActivity"
             android:configChanges="orientation|screenSize|keyboard">
         </activity>
     </application>
