+++ MoodSpace/app/src/main/res/layout/activity_login.xml
@@ -26,7 +26,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/app_name"
-            android:textColor="@color/colorTextTitle"
+            android:textColor="@color/colorTitle"
             android:layout_marginTop="10dp"
             android:textSize="30sp"/>
