+++ MoodSpace/app/src/main/res/values/colors.xml
@@ -6,5 +6,5 @@
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorBlack">#000000</color>
     <color name="colorText">#666666</color>
-    <color name="colorTextTitle">#8d5500</color>
+    <color name="colorTitle">#8d5500</color>
 </resources>
